% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{is_testing}
\alias{is_testing}
\title{Detect Testing Context}
\usage{
is_testing()
}
\value{
Logical indicating if in testing mode.
}
\description{
Determines if code is running in a testing context (testthat or sandbox).
}
\keyword{internal}
